idk yet/hardware_monitor/requirements.txt
Flask==2.3.3
psutil==5.9.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
flask-jwt-extended==4.5.3
flask-limiter==3.5.0
flask-talisman==1.1.0
flask-compress==1.18
flask-caching==2.3.1
prometheus-flask-exporter==0.23.2
flask-httpauth==4.8.0
pytest==8.4.1
pytest-flask==1.3.0
requests==2.32.4 